Fix the `vcs.present` function loop in fish 3
In previous versions, somehow setting the same variable with `set -l`
inside a `while` loop changed the value of the variable with the same name
outside of the scope of the `while`, but in fish 3 the `-l` flag made
the inner variable truly local to the inner scope, while keeping the
outer scope variable intact, thus resulting in an infinite loop.
(might be related to changes introduced while fixing
https://github.com/fish-shell/fish-shell/issues/4820)

`vcs.present` functions for both Mercurial (hg) and git were affected by
this change, resulting in infinite loops in fish 3 while initializing
VCS-aware themes (including the "default" theme).